Moving middleware test folder under development
