--- lib/spec/middleware/HTTPMessageHandler.d.ts
@@ -1,7 +1,0 @@
-/**
- * -------------------------------------------------------------------------------------------
- * Copyright (c) Microsoft Corporation.  All Rights Reserved.  Licensed under the MIT License.
- * See License in the project root for license information.
- * -------------------------------------------------------------------------------------------
- */
-import "isomorphic-fetch";
