+++ package.json
@@ -33,11 +33,12 @@
     "web:js": "node browserify-with-dependencies.js > lib/graph-js-sdk-web.js && uglifyjs ./lib/graph-js-sdk-web.js --output ./lib/graph-js-sdk-web.js",
     "core:js": "node browserify.js > lib/graph-js-sdk-core.js && uglifyjs ./lib/graph-js-sdk-core.js --output ./lib/graph-js-sdk-core.js",
     "build": "npm run compile && npm run web:js && npm run core:js",
-    "test": "mocha lib/spec/content && mocha lib/spec/core && mocha lib/spec/middleware && mocha lib/spec/tasks",
+    "test": "mocha lib/spec/content && mocha lib/spec/core && mocha lib/spec/tasks",
     "test:content": "tsc -p spec && mocha spec/content",
     "test:core": "tsc -p spec && mocha spec/core",
-    "test:middleware": "tsc -p spec && mocha spec/middleware",
     "test:tasks": "tsc -p spec && mocha spec/tasks",
+    "test:development": "tsc -p spec && mocha spec/development/middleware && mocha spec/development/workload",
+    "test:middleware": "tsc -p spec && mocha spec/development/middleware",
     "test:workload": "tsc -p spec && mocha spec/development/workload"
   },
   "repository": {
