+++ spec/development/middleware/HTTPMessageHandler.ts
@@ -7,8 +7,8 @@
 
 import { assert } from "chai";
 import fetchMock = require("fetch-mock");
-import { HTTPMessageHandler } from "../../src/middleware/HTTPMessageHandler";
-import { Context } from "../../src/IContext";
+import { HTTPMessageHandler } from "../../../src/middleware/HTTPMessageHandler";
+import { Context } from "../../../src/IContext";
 import "isomorphic-fetch";
 
 describe("HTTPMessageHandler.ts", () => {
